.github/releases/requirements.txt
semver == 2.13.0
lxml == 4.8.0
pyyaml == 6.0
ruamel.yaml == 0.17.21
ruamel.yaml.clib == 0.2.6